requirements.txt
flask
flask_login
flask_migrate
flask_wtf
flask_sqlalchemy
flask_bcrypt
gunicorn
email_validator
cryptography